Move fmt_count and fmt_size to a new util module.
